Update delete function to receive obj as argument

+++ models/base_model.py
@@ -59,7 +59,8 @@ class BaseModel:
         dictionary.pop('_sa_instance_state', None)
         return dictionary
 
-    def delete(self):
+    def delete(self, obj=None):
         """Delete the instance storage."""
         from models import storage
-        storage.delete(self)
+        if obj is not None:
+            storage.delete(self)
